+++ components/Endeavors/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Endeavors'
+export { default } from './EndeavorsComp'
