--- data/endeavors/endeavors.ts
@@ -1,17 +1,0 @@
-export const endeavorsList = [
-  {
-    img: 'static/images/banner.png',
-    title: 'cretu.dev',
-    description: 'My personal corner of the internet.',
-    link: 'https://github.com/cristicretu/cretu.dev',
-    timeline: 'August 2021 - May 2022',
-  },
-  {
-    img: '/static/images/endeavors/streak.png',
-    title: 'Keep the streak',
-    description:
-      'GitHub Action that notifies the user when he is close to lose the commit streak.',
-    link: '',
-    timeline: 'April 2022',
-  },
-]
