+++ data/endeavors/endeavorsItems.ts
@@ -1,0 +1,40 @@
+export const endeavorsList = [
+  {
+    img: '/static/images/banner.png',
+    title: 'cretu.dev',
+    description: 'My personal corner of the internet.',
+    link: 'https://github.com/cristicretu/cretu.dev',
+    timeline: 'August 2021 - May 2022',
+  },
+  {
+    img: '/static/images/endeavors/esentza.png',
+    title: 'Esentza Robotics',
+    description:
+      'This is where we make robots under the First Tech Challenge competition.',
+    link: 'https://esentza.cretu.dev',
+    timeline: 'April 2022',
+  },
+  {
+    img: '/static/images/endeavors/streak.png',
+    title: 'Keep the streak',
+    description:
+      'GitHub Action that notifies the user when he is close to lose the commit streak.',
+    link: 'https://github.com/cristicretu/keep-the-streak',
+    timeline: 'April 2022',
+  },
+  {
+    img: '/static/images/endeavors/landmarks.png',
+    title: 'Landmarks',
+    description:
+      'GitHub Action that notifies the user when he is close to lose the commit streak.',
+    link: 'https://landmarks.ro',
+    timeline: '2021 - March 2022',
+  },
+  {
+    img: '/static/images/endeavors/wordle.png',
+    title: 'Romanian Wordle',
+    description: 'Weekend project where I built a romanian wordle version.',
+    link: 'https://github.com/cristicretu/wordle',
+    timeline: 'January 2022',
+  },
+]
