+++ public/feed.xml
@@ -4,7 +4,7 @@
         <description><![CDATA[Cristian Crețu]]></description>
         <link>https://cretu.dev</link>
         <generator>RSS for Node</generator>
-        <lastBuildDate>Sun, 29 May 2022 14:48:48 GMT</lastBuildDate>
+        <lastBuildDate>Sun, 29 May 2022 16:13:24 GMT</lastBuildDate>
         <atom:link href="https://cretu.dev/feed.xml" rel="self" type="application/rss+xml"/>
         <item>
             <title><![CDATA[2021 in Review]]></title>
